+++ lib/buttons.dart
@@ -1,1 +1,2 @@
-export 'buttons/rect_button_component.dart';
+export 'buttons/rectangle_button_component.dart';
+export 'buttons/circle_button_component.dart';