+++ lib/inputs.dart
@@ -1,1 +1,2 @@
 export 'inputs/text_field_component.dart';
+export 'inputs/toggle_component.dart';